constants to svelteKit basics & codewrapper component
