+++ src/routes/sveltekit-basics/shared-modules/constants.ts
@@ -1,0 +1,11 @@
+export const HEADER = {
+	SHARED_MODULES: 'shared-modules.svelte'
+};
+
+export const CODE = {
+	SHARING_MODULES: `
+	<script>
+        import CodeBlock from "$lib/components/ui/code-block/code-block.svelte";
+    </script>
+	`
+};
